requirements.txt
matplotlib>=3.5.0
Pillow>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
tqdm>=4.65.0

# run "/Applications/Python\ 3.11/Install\ Certificates.command" in terminal if SSL errors occur on MacOS